% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_expex.R
\name{excel_exporter_server}
\alias{excel_exporter_server}
\title{Server Export excel to symbol separated file}
\usage{
excel_exporter_server(id)
}
\arguments{
\item{id}{session id}
}
\description{
Server Export excel to symbol separated file
}
